fix:fix some bugs of the git

+++ code/backend/weibo-service/src/main/java/com/ilife/weiboservice/dao/UserDao.java
@@ -1,10 +1,10 @@
 package com.ilife.weiboservice.dao;
 
-import com.ilife.weiboservice.entity.User;
+        import com.ilife.weiboservice.entity.User;
 
 
-public interface UserDao{
-    User findAllById(Integer id);
+public interface UserDao {
+    User findAllById(Long id);
 
     User findByNickname(String nickname);
